feat: 🎈 (api/modules/cors) ファイルを一つにする修正
